# openapi.yaml
openapi: 3.0.3
info:
  title: Flask CRUD
  version: 1.0.0
...
paths:
  /attack_types:
    post:
      summary: Create an attack type
      tags:
        - AttackTypes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
              properties:
                type:
                  type: string
                  example: Phishing
      responses:
        "201":
          description: Attack type created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttackType"
        "400":
          description: Invalid type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      summary: List all attack types
      tags:
        - AttackTypes
      responses:
        "200":
          description: List of attack types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttackType"
  /attack_types/{id}:
    get:
      summary: Get an attack type by ID
      tags:
        - AttackTypes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: Attack type found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttackType"
        "404":
          description: Attack type not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      summary: Update an attack type
      tags:
        - AttackTypes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
              properties:
                type:
                  type: string
                  example: Ransomware
      responses:
        "200":
          description: Attack type updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttackType"
        "400":
          description: Invalid type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Attack type not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete an attack type
      tags:
        - AttackTypes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "204":
          description: Attack type deleted
        "404":
          description: Attack type not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Attack type cannot be deleted due to foreign key constraint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    AttackType:
      type: object
      properties:
        id:
          type: integer
          example: 1
        type:
          type: string
          example: Phishing
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Not found